edited userManager.js

+++ src/components/settings/service/userManager.js
@@ -65,19 +65,32 @@ export const userManager = {
         console.error('Error:', error);
       });
   },
-  deleteUser: function (id, queryClient, dispatch) {
-    fetch('http://localhost:4000/user/' + id, {
-      method: 'DELETE'
-    })
-      .then((response) => response.json())
-      .then(() => {
-        queryClient.invalidateQueries();
-        showNotification('User deleted ', 'success', dispatch);
-      })
-      .catch((error) => {
-        showNotification('Error deleting user! Please try again', 'error', dispatch);
-        console.error('Error:', error);
+  deleteUser: async function (id, queryClient, dispatch) {
+    try {
+      const userToken = localStorage.getItem('userToken');
+      if (!userToken) {
+        throw new Error('User token is missing');
+      }
+
+      const response = await fetch(`http://localhost:8080/api/user/delete/${id}`, {
+        method: 'DELETE',
+        headers: {
+          Authorization: `Bearer ${userToken}`
+        }
       });
+
+      if (response.ok) {
+        queryClient.invalidateQueries();
+        showNotification('User deleted successfully.', 'info', dispatch);
+      } else {
+        const errorData = await response.text();
+        console.error('Error:', errorData);
+        showNotification('Failed to delete user. Please try again.', 'error', dispatch);
+      }
+    } catch (error) {
+      console.error('Network error:', error.message);
+      showNotification('Network error: Unable to delete user.', 'error', dispatch);
+    }
   },
   updateUser: async function (data, queryClient, dispatch) {
     try {
@@ -118,6 +131,35 @@ export const userManager = {
     } catch (error) {
       console.log(error);
       return false;
+    }
+  },
+
+  manageUser: async function (id, action, queryClient, dispatch) {
+    try {
+      const userToken = localStorage.getItem('userToken');
+      if (!userToken) {
+        throw new Error('User token is missing');
+      }
+
+      const response = await fetch(`http://localhost:8080/api/user/manageUser/${id}/${action}`, {
+        method: 'PUT',
+        headers: {
+          'Content-Type': 'application/json',
+          Authorization: `Bearer ${userToken}`
+        }
+      });
+
+      if (response.ok) {
+        queryClient.invalidateQueries();
+        showNotification(`User updated successfully`, 'success', dispatch);
+      } else {
+        const errorData = await response.text();
+        console.error('Error:', errorData);
+        showNotification(`Failed to update user. ${errorData}.`, 'error', dispatch);
+      }
+    } catch (error) {
+      console.error('Network error:', error.message);
+      showNotification('Network error: Unable to update user.', 'error', dispatch);
     }
   }
 };
